% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{ConcatSamples}
\alias{ConcatSamples}
\title{Take directory of CL files and concatenate, sort, and re-ID them. This preserves all values from the original CL}
\usage{
ConcatSamples(Indirectory, Outdirectory)
}
\arguments{
\item{Indirectory}{Specify directory of CL files and only CL files There is no Type checking, it will just fail if not in the correct format.}

\item{Outdirectory}{Specify an directory including the desired name of the output file to write the concatenated list.}
}
\value{
A Query List (QL) style output using the origional CL information
}
\description{
If a user has many samples they may want to run each sample by itself first, to remove redundancy within a sample, and then combine all sample CLs to do an dataset-wide analysis.
I.e. single sample QL -> Single sample CL -> ConcatSamples (multisample) -> ConsensusToQuery.
}
\examples{
ConcatSamples(~user/R/BunchofCLs, ~User/R/concatOut/QL.out.tsv)
}
