% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{ConsensusToQuery}
\alias{ConsensusToQuery}
\title{Convert a consensus style list to a query style list}
\usage{
ConsensusToQuery(fileIn, fileOut)
}
\arguments{
\item{fileIn}{Consensus List (with path) , in the CL format. Typically the output of a previous run}

\item{fileOut}{Query List (with path), in the QL format.}
}
\description{
! re-IDs the input !
}
\examples{
ConsensusToQuery("~/user/R/ConsensusList.tsv", "~/user/R/QueryListOut.tsv")
}
