% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hunt.R
\name{Hunt}
\alias{Hunt}
\title{Run SVCROWS in "Hunt" mode}
\usage{
Hunt(
  InputQueryList,
  FeatureList,
  OutputDirectory,
  BPfactor = TRUE,
  DefaultSizes = FALSE,
  xs = NA,
  xl = NA,
  y1s = NA,
  y1l = NA,
  y2s = NA,
  y2l = NA
)
}
\arguments{
\item{InputQueryList}{Query List in the designated format (see user manual)}

\item{FeatureList}{Similar to the Query List, with slight differences. Contains a list of all features the user wants to query the Query List against. (see user manual)}

\item{OutputDirectory}{Directory to write output files (see user manual)}

\item{BPfactor}{Boolean. By Default TRUE, uses breakpoints as a secondary piece of information to call RO overlapping regions. Regardless, matching regions will be counted in the FCL.}

\item{DefaultSizes}{Boolean. By default False, which uses the other 6 inputs given by the user. If set to True, SVCROWS will use the 2nd and 4th quartiles to generate values for the six variables (see user manual for details).}

\item{xs}{Int. Small SV size cutoff}

\item{xl}{Int. Large SV size cutoff}

\item{y1s}{Int. Small SV BP region size}

\item{y1l}{Int. Large SV BP region size}

\item{y2s}{Int. Small SV RO Threshold}

\item{y2l}{Int. Small SV RO Threshold}
}
\description{
This mode uses the framework of overlap comparison detailed for "scavange" mode, but applies it to overlapping features in the reference genome. For example, if you want to know what set of genes are captured in your Query List, you can input a 'Feature list, which contains information about all of the genes in the genome. Consider using for intergenic regions, Transcriptional start sites, enhancers, etc.
}
\examples{
Hunt("~/user/R/SVCROWSin", "~/user/R/FeatureList", "~/user/R/SVCROWSout", TRUE, FALSE, 5000, 25000, 500, 2500, 50, 80)
Hunt("~/user/R/SVCROWSin", "~/user/R/FeatureList", "~/user/R/SVCROWSout", TRUE, TRUE)
}
