% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scavenge.R
\name{Scavenge}
\alias{Scavenge}
\title{Run SVCROWS in "Scavenge" mode}
\usage{
Scavenge(
  InputQueryList,
  OutputDirectory,
  ExpandRORegion = FALSE,
  BPfactor = TRUE,
  DefaultSizes = FALSE,
  xs = NA,
  xl = NA,
  y1s = NA,
  y1l = NA,
  y2s = NA,
  y2l = NA
)
}
\arguments{
\item{InputQueryList}{Query List in the designated format (see user manual)}

\item{OutputDirectory}{Directory to write output files (see user manual)}

\item{ExpandRORegion}{If TRUE, When entries match, use the minimum and maximum breakpoints of either to define a new region (which will be used for all future comparisons). When FALSE, original bounds are kept}

\item{BPfactor}{Boolean. By Default TRUE, uses breakpoints as a secondary piece of information to call RO overlapping regions. Regardless, matching regions will be counted in the FCL.}

\item{DefaultSizes}{Boolean. By default False, which uses the other 6 inputs given by the user. If set to True, SVCROWS will use the 2nd and 4th quartiles to generate values for the six variables (see user manual for details).}

\item{xs}{Int. Small SV size cutoff}

\item{xl}{Int. Large SV size cutoff}

\item{y1s}{Int. Small SV BP region size}

\item{y1l}{Int. Large SV BP region size}

\item{y2s}{Int. Small SV RO Threshold}

\item{y2l}{Int. Small SV RO Threshold}
}
\description{
#' @description This is the main function of SVCROWS. This is a program uses reciprocal overlap (RO) to determine if two regions have a significant level of overlap. This program allows for user input to determine the stringency of the comparisons, while also giving options to use other pieces of evidence to support RO calls. In principle, the program uses the concept of "Weighted Sizes" to weight RO stringency.
}
\examples{
Hunt("~/user/R/SVCROWSin", "~/user/R/SVCROWSout", FALSE, TRUE, FALSE, 5000, 25000, 500, 2500, 50, 80)
Hunt("~/user/R/SVCROWSin", "~/user/R/SVCROWSout", TRUE, TRUE, TRUE)
}
