% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{SVCSummarize}
\alias{SVCSummarize}
\title{SVCSummarize}
\usage{
SVCSummarize(OutputDirectory, OutputFile)
}
\arguments{
\item{OutputDirectory}{Directory where the output files are being written to from either Hunt() or Scavenge()}

\item{OutputFile}{What the summary file will be written as}
}
\description{
SVCSummarize
}
\examples{
SVCSummarize("~/R/SVCROWS2/SVCROWSout", "~/R/SVCROWS2/SVCROWSout/summary.tsv")
}
